Change atoi to atol, causing obvious issues on what needs to be int64's (#7466)
The system stats plugin is currently using atoi and ints. These need to be longs to match int64

+++ plugins/experimental/system_stats/system_stats.c
@@ -112,7 +112,7 @@ getFile(const char *filename, char *buffer, int bufferSize)
 }
 
 static void
-statSet(const char *name, int value, TSMutex stat_creation_mutex)
+statSet(const char *name, long value, TSMutex stat_creation_mutex)
 {
   int stat_id = statAdd(name, TS_RECORDDATATYPE_INT, stat_creation_mutex);
   if (stat_id != TS_ERROR) {
@@ -153,7 +153,7 @@ setNetStat(TSMutex stat_creation_mutex, const char *interface, const char *entry
   if (getFile(&sysfs_name[0], &data[0], sizeof(data)) < 0) {
     TSDebug(DEBUG_TAG, "Error reading file %s", sysfs_name);
   } else {
-    statSet(stat_name, atoi(data), stat_creation_mutex);
+    statSet(stat_name, atol(data), stat_creation_mutex);
   }
 }
